More renaming to be explicit about Chat memory repository
