+++ memory/repository/spring-ai-model-chat-memory-repository-jdbc/src/main/java/org/springframework/ai/chat/memory/repository/jdbc/HsqldbChatMemoryRepositoryDialect.java
@@ -17,31 +17,28 @@
 package org.springframework.ai.chat.memory.repository.jdbc;
 
 /**
- * Dialect for Postgres.
- *
- * @author Mark Pollack
- * @since 1.0.0
+ * HSQLDB-specific SQL dialect for chat memory repository.
  */
-public class PostgresChatMemoryDialect implements JdbcChatMemoryDialect {
+public class HsqldbChatMemoryRepositoryDialect implements JdbcChatMemoryRepositoryDialect {
 
 	@Override
 	public String getSelectMessagesSql() {
-		return "SELECT content, type FROM SPRING_AI_CHAT_MEMORY WHERE conversation_id = ? ORDER BY \"timestamp\"";
+		return "SELECT content, type FROM SPRING_AI_CHAT_MEMORY WHERE conversation_id = ? ORDER BY timestamp ASC";
 	}
 
 	@Override
 	public String getInsertMessageSql() {
-		return "INSERT INTO SPRING_AI_CHAT_MEMORY (conversation_id, content, type, \"timestamp\") VALUES (?, ?, ?, ?)";
-	}
-
-	@Override
-	public String getSelectConversationIdsSql() {
-		return "SELECT DISTINCT conversation_id FROM SPRING_AI_CHAT_MEMORY";
+		return "INSERT INTO SPRING_AI_CHAT_MEMORY (conversation_id, content, type, timestamp) VALUES (?, ?, ?, ?)";
 	}
 
 	@Override
 	public String getDeleteMessagesSql() {
 		return "DELETE FROM SPRING_AI_CHAT_MEMORY WHERE conversation_id = ?";
+	}
+
+	@Override
+	public String getSelectConversationIdsSql() {
+		return "SELECT DISTINCT conversation_id FROM SPRING_AI_CHAT_MEMORY";
 	}
 
 }
