+++ memory/repository/spring-ai-model-chat-memory-repository-jdbc/src/main/java/org/springframework/ai/chat/memory/repository/jdbc/SqlServerChatMemoryRepositoryDialect.java
@@ -22,7 +22,7 @@ package org.springframework.ai.chat.memory.repository.jdbc;
  * @author Mark Pollack
  * @since 1.0.0
  */
-public class SqlServerChatMemoryDialect implements JdbcChatMemoryDialect {
+public class SqlServerChatMemoryRepositoryDialect implements JdbcChatMemoryRepositoryDialect {
 
 	@Override
 	public String getSelectMessagesSql() {
