+++ memory/repository/spring-ai-model-chat-memory-repository-jdbc/src/test/java/org/springframework/ai/chat/memory/repository/jdbc/JdbcChatMemoryRepositoryPostgresqlIT.java
@@ -164,7 +164,7 @@ class JdbcChatMemoryRepositoryPostgresqlIT {
 		ChatMemoryRepository chatMemoryRepository(JdbcTemplate jdbcTemplate, DataSource dataSource) {
 			return JdbcChatMemoryRepository.builder()
 				.jdbcTemplate(jdbcTemplate)
-				.dialect(JdbcChatMemoryDialect.from(dataSource))
+				.dialect(JdbcChatMemoryRepositoryDialect.from(dataSource))
 				.build();
 		}
